# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * StreamFlix API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserProfileRole = typeof UserProfileRole[keyof typeof UserProfileRole];


export const UserProfileRole = {
  user: 'user',
  admin: 'admin',
  moderator: 'moderator',
} as const;

export interface UserProfile {
  id: string;
  clerkId: string;
  email: string;
  /** @nullable */
  username?: string | null;
  role: UserProfileRole;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}

export type UserRoleUpdateRole = typeof UserRoleUpdateRole[keyof typeof UserRoleUpdateRole];


export const UserRoleUpdateRole = {
  user: 'user',
  admin: 'admin',
  moderator: 'moderator',
} as const;

export interface UserRoleUpdate {
  role: UserRoleUpdateRole;
}

export interface Category {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export interface CategoryInput {
  name: string;
  description?: string;
}

export interface CategoryUpdate {
  name?: string;
  description?: string;
}

export type ContentContentType = typeof ContentContentType[keyof typeof ContentContentType];


export const ContentContentType = {
  movie: 'movie',
  series: 'series',
} as const;

export interface Content {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  posterUrl?: string | null;
  /** @nullable */
  backdropUrl?: string | null;
  /** @nullable */
  videoUrl?: string | null;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  genre?: string | null;
  /** @nullable */
  releaseYear?: number | null;
  /** @nullable */
  durationMinutes?: number | null;
  contentType: ContentContentType;
  isFeatured: boolean;
  /** @nullable */
  averageRating?: number | null;
  viewCount: number;
  createdAt: string;
}

export type ContentDetailContentType = typeof ContentDetailContentType[keyof typeof ContentDetailContentType];


export const ContentDetailContentType = {
  movie: 'movie',
  series: 'series',
} as const;

export interface Episode {
  id: number;
  seasonId: number;
  episodeNumber: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  videoUrl?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  durationMinutes?: number | null;
  createdAt: string;
}

export interface SeasonWithEpisodes {
  id: number;
  contentId: number;
  seasonNumber: number;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  description?: string | null;
  createdAt: string;
  episodes: Episode[];
}

export interface ContentDetail {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  posterUrl?: string | null;
  /** @nullable */
  backdropUrl?: string | null;
  /** @nullable */
  videoUrl?: string | null;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  genre?: string | null;
  /** @nullable */
  releaseYear?: number | null;
  /** @nullable */
  durationMinutes?: number | null;
  contentType: ContentDetailContentType;
  isFeatured: boolean;
  /** @nullable */
  averageRating?: number | null;
  viewCount: number;
  createdAt: string;
  seasons: SeasonWithEpisodes[];
  isFavorite: boolean;
  /** @nullable */
  userRating?: number | null;
  /** @nullable */
  watchProgress?: number | null;
}

export interface ContentListResponse {
  items: Content[];
  total: number;
}

export type ContentInputContentType = typeof ContentInputContentType[keyof typeof ContentInputContentType];


export const ContentInputContentType = {
  movie: 'movie',
  series: 'series',
} as const;

export interface ContentInput {
  title: string;
  description?: string;
  posterUrl?: string;
  backdropUrl?: string;
  videoUrl?: string;
  categoryId?: number;
  genre?: string;
  releaseYear?: number;
  durationMinutes?: number;
  contentType: ContentInputContentType;
  isFeatured?: boolean;
}

export type ContentUpdateContentType = typeof ContentUpdateContentType[keyof typeof ContentUpdateContentType];


export const ContentUpdateContentType = {
  movie: 'movie',
  series: 'series',
} as const;

export interface ContentUpdate {
  title?: string;
  description?: string;
  posterUrl?: string;
  backdropUrl?: string;
  videoUrl?: string;
  categoryId?: number;
  genre?: string;
  releaseYear?: number;
  durationMinutes?: number;
  contentType?: ContentUpdateContentType;
  isFeatured?: boolean;
}

export interface Season {
  id: number;
  contentId: number;
  seasonNumber: number;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export interface SeasonInput {
  seasonNumber: number;
  title?: string;
  description?: string;
}

export interface SeasonUpdate {
  seasonNumber?: number;
  title?: string;
  description?: string;
}

export interface EpisodeInput {
  episodeNumber: number;
  title: string;
  description?: string;
  videoUrl?: string;
  thumbnailUrl?: string;
  durationMinutes?: number;
}

export interface EpisodeUpdate {
  episodeNumber?: number;
  title?: string;
  description?: string;
  videoUrl?: string;
  thumbnailUrl?: string;
  durationMinutes?: number;
}

export interface WatchHistoryEntry {
  id: number;
  userId: string;
  contentId: number;
  /** @nullable */
  episodeId?: number | null;
  progressSeconds: number;
  completed: boolean;
  watchedAt: string;
  content?: Content;
}

export interface WatchProgressInput {
  progressSeconds: number;
  episodeId?: number;
  completed?: boolean;
}

export interface Rating {
  id: number;
  userId: string;
  contentId: number;
  score: number;
  createdAt: string;
}

export interface RatingInput {
  score: number;
}

export interface Comment {
  id: number;
  userId: string;
  contentId: number;
  text: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}

export interface CommentInput {
  text: string;
}

export interface AdminStats {
  totalUsers: number;
  totalContent: number;
  totalMovies: number;
  totalSeries: number;
  totalCategories: number;
  totalWatchEvents: number;
  totalComments: number;
  topContent: Content[];
  recentActivity: WatchHistoryEntry[];
}

export interface UploadUrlRequest {
  filename: string;
  contentType: string;
}

export interface UploadUrlResult {
  uploadUrl: string;
  objectPath: string;
}

export type ListContentParams = {
categoryId?: number;
genre?: string;
type?: ListContentType;
limit?: number;
offset?: number;
};

export type ListContentType = typeof ListContentType[keyof typeof ListContentType];


export const ListContentType = {
  movie: 'movie',
  series: 'series',
} as const;

export type GetLatestContentParams = {
limit?: number;
};

export type GetPopularContentParams = {
limit?: number;
};

export type SearchContentParams = {
q: string;
type?: SearchContentType;
};

export type SearchContentType = typeof SearchContentType[keyof typeof SearchContentType];


export const SearchContentType = {
  movie: 'movie',
  series: 'series',
} as const;