# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: StreamFlix API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get category by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # CONTENT (Movies & Series)
...
  /admin/users/{userId}/role:
    patch:
      operationId: updateUserRole
      tags: [admin]
      summary: Update user role (admin)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRoleUpdate"
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /admin/users/{userId}:
    delete:
      operationId: deleteUser
      tags: [admin]
      summary: Delete user (admin)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # STORAGE
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        clerkId:
          type: string
        email:
          type: string
        username:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin, moderator]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - clerkId
        - email
        - role
        - createdAt
    UserRoleUpdate:
      type: object
      properties:
        role:
          type: string
          enum: [user, admin, moderator]
      required:
        - role
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - createdAt
    CategoryInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
      required:
        - name
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
...